Skipping existing files

+++ data-collection/Extract_Text.py
@@ -4,9 +4,7 @@ import re
 
 def get_text_from_html(html):
     matches = re.findall(r'"text":"(.*?)"', html)
-
     matches = list(dict.fromkeys(matches))
-
     text = ""
     for match in matches:
         text += match + "\n"
@@ -26,11 +24,8 @@ def get_output_path(path):
     return [f"data/articles/{topic}/{year}/month{month}/day{day}/", f"{index}.txt"]
 
 
-# start variables
 start_year = 2020
-# start_year = config.get_input_number("Input Start Year: ")
 amount_years = 2
-# amount_years = config.get_input_number("Input amount of years: ")
 
 topics = ["politics", "world", "opinion"]
 for topic in topics:
@@ -49,13 +44,17 @@ for topic in topics:
                     if file.endswith('.txt'):
                         files.append(os.path.join(files_path, file))
             for file in files:
-                with open(file, 'r', encoding='utf-8') as f:
-                    html_content = f.read()
-
                 output = get_output_path(file)
                 os.makedirs(output[0], exist_ok=True)
                 output_file_path = output[0] + output[1]
 
+                if os.path.exists(output_file_path):
+                    print(
+                        f"File {output_file_path} already exists. Skipping...")
+                    continue
+
+                with open(file, 'r', encoding='utf-8') as f:
+                    html_content = f.read()
                 article_text = get_text_from_html(html_content)
 
                 with open(output_file_path, "w", encoding="utf-8") as f:
